# api_service/static/openapi.yml
openapi: 3.0.0
info:
  title: User Authentication API
  description: API for user registration, authentication, and profile management
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Proxy Server
  - url: http://localhost:500
    description: User Service
paths:
  /api/v1/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
                email:
                  type: string
                nickname:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                phone_number:
                  type: string
              required:
                - login
                - password
                - email
                - nickname
                - date_of_birth
                - phone_number
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request (missing fields, invalid data, etc.)
  /api/v1/login:
    post:
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                login:
                  type: string
                password:
                  type: string
              required:
                - login
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
        '400':
          description: Invalid login or password
  /api/v1/profile/{nickname}:
    put:
      summary: Update user profile
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                phone_number:
                  type: string
              required:
                - nickname
                - date_of_birth
                - phone_number
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request (missing fields, invalid data, etc.)
        '404':
          description: User not found
    get:
      summary: Get user profile
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  login:
                    type: string
                  email:
                    type: string
                  nickname:
                    type: string
                  date_of_birth:
                    type: string
                    format: date
                  phone_number:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '404':
          description: User not found